dumping comparison result to json file and moving some functions to new module
